Fix grammar in README.md
This change makes "controller" plural so that it agrees with the plural subject "action creators."  Because this plural form continued into the second sentence with "They are simply objects..." I also changed "manages" to the plural form. Likewise, in the third sentence, I changed "It allows" into "They allow" so that these three sentences are all consistent grammatically and agree with the plural form of the subject "action creators". Nothing else is changed.